// src/Store/Releases/types.ts
enum ActionTypeKeys {
    GET_RELEASE_DEFINITIONS_REQUEST = "GET_RELEASE_DEFINITIONS_REQUEST",
    GET_RELEASE_DEFINITIONS_SUCCESS = "GET_RELEASE_DEFINITIONS_SUCCESS",
    GET_RELEASE_DEFINITIONS_FAILURE = "GET_RELEASE_DEFINITIONS_FAILURE",
    SET_CURRENT_RELEASE_DEFINITION = "SET_CURRENT_RELEASE_DEFINITION",
    GET_RELEASES_REQUEST = "GET_RELEASES_REQUEST",
    GET_RELEASES_FAILURE = "GET_RELEASES_FAILURE",
    GET_RELEASES_SUCCESS = "GET_RELEASES_SUCCESS",
    SET_CURRENT_RELEASE="SET_CURRENT_RELEASE",
    SET_CURRENT_ENVIRONMENT ="SET_CURRENT_ENVIRONMENT",
    GET_WORKITEMS_REQUEST="GET_WORKITEMS_REQUEST",
    GET_WORKITEMS_SUCCESS="GET_WORKITEMS_SUCCESS",
    GET_WORKITEMS_FAILURE="GET_WORKITEMS_FAILURE"
}
export default ActionTypeKeys;